chore: moving reactor to correct package, adding new reactor to get the current room set for an insight
