+++ src/prerna/reactor/insights/SetRoomForInsightReactor.java
@@ -25,7 +25,7 @@
  * 	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * 	GNU General Public License for more details.
  *******************************************************************************/
-package prerna.reactor.project;
+package prerna.reactor.insights;
 
 import java.util.List;
 import java.util.Map;
